fix: slide→text
