+++ frontend/src/components/DraggableSlideImage.tsx
@@ -1,53 +1,186 @@
 import type React from 'react'
-import { EditorContent } from '@tiptap/react'
-import styles from './DraggableTextBox.module.css'
-import { useEffect, useCallback } from 'react'
-import { useDrag } from '@/hooks/useDrag'
-import { useResize } from '@/hooks/useResize'
-import type { TextBox } from '@/types/Slide'
-import type { ResizeDivs } from '@/types/DraggableTextBox'
+import styles from './DraggableSlideImage.module.css'
+import { useState, useCallback, useEffect } from 'react'
+import type { SlideImage } from '@/types/Slide'
+import resizeStyles from './ResizeHandles.module.css'
 
 type Props = {
-  textbox: TextBox
+  image: SlideImage
 }
 
+import type {
+  Direction,
+  ResizeOptions,
+  ResizeDivs,
+} from '@/types/DraggableTextBox'
+
+const handleResize = (
+  e: MouseEvent,
+  resizeStart: { x: number; y: number },
+  size: { width: number; height: number },
+  position: { x: number; y: number },
+  direction: Direction,
+): ResizeOptions => {
+  const deltaX = e.clientX - resizeStart.x
+  const deltaY = e.clientY - resizeStart.y
+
+  const options: { [key in Direction]: ResizeOptions } = {
+    north: {
+      width: size.width,
+      height: size.height - deltaY,
+      x: position.x,
+      y: position.y + deltaY,
+    },
+    northEast: {
+      width: size.width + deltaX,
+      height: size.height - deltaY,
+      x: position.x,
+      y: position.y + deltaY,
+    },
+    east: {
+      width: size.width + deltaX,
+      height: size.height,
+      x: position.x,
+      y: position.y,
+    },
+    southEast: {
+      width: size.width + deltaX,
+      height: size.height + deltaY,
+      x: position.x,
+      y: position.y,
+    },
+    south: {
+      width: size.width,
+      height: size.height + deltaY,
+      x: position.x,
+      y: position.y,
+    },
+    southWest: {
+      width: size.width - deltaX,
+      height: size.height + deltaY,
+      x: position.x + deltaX,
+      y: position.y,
+    },
+    west: {
+      width: size.width - deltaX,
+      height: size.height,
+      x: position.x + deltaX,
+      y: position.y,
+    },
+    northWest: {
+      width: size.width - deltaX,
+      height: size.height - deltaY,
+      x: position.x + deltaX,
+      y: position.y + deltaY,
+    },
+    default: {
+      width: size.width,
+      height: size.height,
+      x: position.x,
+      y: position.y,
+    },
+  }
+
+  return options[direction]
+}
 const handleResizeDivs: ResizeDivs[] = [
-  { direction: 'north', className: styles.resizeHandleNorth },
-  { direction: 'northEast', className: styles.resizeHandleNorthEast },
-  { direction: 'east', className: styles.resizeHandleEast },
-  { direction: 'southEast', className: styles.resizeHandleSouthEast },
-  { direction: 'south', className: styles.resizeHandleSouth },
-  { direction: 'southWest', className: styles.resizeHandleSouthWest },
-  { direction: 'west', className: styles.resizeHandleWest },
-  { direction: 'northWest', className: styles.resizeHandleNorthWest },
+  { direction: 'north', className: resizeStyles.resizeHandleNorth },
+  { direction: 'northEast', className: resizeStyles.resizeHandleNorthEast },
+  { direction: 'east', className: resizeStyles.resizeHandleEast },
+  { direction: 'southEast', className: resizeStyles.resizeHandleSouthEast },
+  { direction: 'south', className: resizeStyles.resizeHandleSouth },
+  { direction: 'southWest', className: resizeStyles.resizeHandleSouthWest },
+  { direction: 'west', className: resizeStyles.resizeHandleWest },
+  { direction: 'northWest', className: resizeStyles.resizeHandleNorthWest },
 ]
 
-const DraggableTextBox: React.FC<Props> = ({ textbox }) => {
-  const {
-    position: dragPosition,
-    handleDragMouseDown,
-    handleDragMouseMove,
-    handleDragMouseUp,
-  } = useDrag({ x: textbox.x, y: textbox.y })
-
-  const {
-    isResizing,
-    size,
-    position: resizePosition,
-    handleResizeMouseDown,
-    handleResizeMouseMove,
-    handleResizeMouseUp,
-  } = useResize({ width: textbox.width, height: textbox.height }, dragPosition)
+const DraggableSlideImage: React.FC<Props> = ({ image }) => {
+  const [isDragging, setIsDragging] = useState(false)
+  const [dragStart, setDragStart] = useState<{ x: number; y: number } | null>(
+    null,
+  )
+  const [position, setPosition] = useState({ x: image.x, y: image.y })
+  const [size, setSize] = useState({
+    width: image.width,
+    height: image.height,
+  })
+  const [isResizing, setIsResizing] = useState(false)
+  const [resizeDirection, setResizeDirection] = useState<Direction>('default')
+  const [resizeStart, setResizeStart] = useState<{ x: number; y: number }>({
+    x: size.width,
+    y: size.height,
+  })
 
   const style: React.CSSProperties = {
-    transform: `translate3d(${resizePosition.x}px, ${resizePosition.y}px, 0)`,
+    transform: `translate3d(${position.x}px, ${position.y}px, 0)`,
     position: 'absolute',
     width: `${size.width}px`,
     height: `${size.height}px`,
     boxSizing: 'border-box',
-    outline: textbox.isSelected ? 'solid 1px blue' : 'none',
-    userSelect: 'none',// Prevent text selection(入力の無効化はtiptapにメソッドが存在する為、注意が必要)
+    outline: image.isSelected ? 'solid 1px blue' : 'none',
+    userSelect: 'none', // Prevent text selection(入力の無効化はtiptapにメソッドが存在する為、注意が必要)
   }
+
+  const handleDragMouseDown = useCallback(
+    (e: React.MouseEvent) => {
+      setDragStart({ x: e.clientX - position.x, y: e.clientY - position.y })
+      setIsDragging(true)
+    },
+    [position.x, position.y],
+  )
+
+  const handleDragMouseMove = useCallback(
+    (e: MouseEvent) => {
+      if (!isDragging || !dragStart) return
+      setPosition({
+        x: e.clientX - dragStart.x,
+        y: e.clientY - dragStart.y,
+      })
+    },
+    [isDragging, dragStart],
+  )
+
+  const handleDragMouseUp = useCallback(() => {
+    if (!isDragging) return
+    setIsDragging(false)
+    setDragStart(null)
+  }, [isDragging])
+
+  const handleResizeMouseDown = useCallback(
+    (e: React.MouseEvent, direction: Direction) => {
+      e.stopPropagation()
+      setResizeDirection(direction)
+      setIsResizing(true)
+      setResizeStart({ x: e.clientX, y: e.clientY })
+    },
+    [],
+  )
+  const handleResizeMouseMove = useCallback(
+    (e: MouseEvent) => {
+      if (!isResizing) return
+      const newOptions = handleResize(
+        e,
+        resizeStart,
+        size,
+        position,
+        resizeDirection,
+      )
+      if (newOptions.width < -10 || newOptions.height < -10) return
+
+      setPosition({ x: newOptions.x, y: newOptions.y })
+      setSize({
+        width: newOptions.width,
+        height: newOptions.height,
+      })
+      setResizeStart({ x: e.clientX, y: e.clientY })
+    },
+    [isResizing, resizeStart, size, position, resizeDirection],
+  )
+  const handleResizeMouseUp = useCallback(() => {
+    if (!isResizing) return
+    setIsResizing(false)
+    setResizeDirection('default')
+  }, [isResizing])
 
   const handleMouseMove = useCallback(
     (e: MouseEvent) => {
@@ -56,7 +189,6 @@ const DraggableTextBox: React.FC<Props> = ({ textbox }) => {
     },
     [handleDragMouseMove, handleResizeMouseMove],
   )
-
   const handleMouseUp = useCallback(() => {
     handleDragMouseUp()
     handleResizeMouseUp()
@@ -69,31 +201,39 @@ const DraggableTextBox: React.FC<Props> = ({ textbox }) => {
       document.removeEventListener('mousemove', handleMouseMove)
       document.removeEventListener('mouseup', handleMouseUp)
     }
-  }, [handleMouseMove, handleMouseUp])
+  }, [
+    handleDragMouseMove,
+    handleDragMouseUp,
+    handleMouseMove,
+    handleMouseUp,
+    handleResizeMouseMove,
+    handleResizeMouseUp,
+  ])
 
   return (
     <div
       style={style}
       onMouseDown={handleDragMouseDown}
-      className={styles.textBox}
+      className={styles.SlideImage}
     >
-      <div onMouseDown={(e) => e.stopPropagation()}>
-        <EditorContent editor={textbox.editor} />
-      </div>
-      <div
-        className={styles.resizeHandles}
-        style={{ display: isResizing ? 'block' : '' }}
-      >
-        {handleResizeDivs.map((div) => (
-          <div
-            key={div.direction}
-            className={div.className}
-            onMouseDown={(e) => handleResizeMouseDown(e, div.direction)}
-          />
-        ))}
-      </div>
+      <img
+        src={image.src}
+        alt="image"
+        onMouseDown={handleDragMouseDown}
+        onDragStart={(e) => e.preventDefault()}
+        width={size.width}
+        height={size.height}
+      />
+
+      {handleResizeDivs.map((div) => (
+        <div
+          key={div.direction}
+          className={div.className}
+          onMouseDown={(e) => handleResizeMouseDown(e, div.direction)}
+        />
+      ))}
     </div>
   )
 }
 
-export default DraggableTextBox
+export default DraggableSlideImage
